+++ cli/src/util/mod.rs
@@ -1,4 +1,5 @@
 pub mod cmd;
+pub mod download;
 pub mod fmt;
 pub mod gen;
 pub mod git;