+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Unreal helper in `rivet init`
+- Installer for the Unreal Engine plugin with `rivet unreal install-plugin` or `rivet init --unreal`
+
 ### Changed
 
 - Renamed `rivet.version.toml` to `rivet.toml`. All changes are backwards compatible.
@@ -43,4 +48,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update dependency: `rivet-api`
 - Upgrade dependency: `tokio 1.27`
 - Removed unnecessary feature flags from `tokio`
-